requirements.txt
requests
typer
python-dotenv
groq
diskcache
dateparser
streamlit
pytest
